HALT work on UI tools for now; here for UI MODULE
Having second thoughts about making the editor so I'm gonna try to just
use Tiled for a while and see how it goes. This means that the UI module
will be halted for now. It's still a useful module for game UI in
general so I'm likely going to continue working on it eventually, but
changing focus for now.